requirements.txt
easydict==1.9
Flask==2.1.0
jieba==0.42.1
numpy==1.19.5
python-dotenv==0.21.0
requests==2.25.1
tensorflow_gpu==2.5.0
